fix quote
