+++ src/types/operations/pools.ts
@@ -1,5 +1,4 @@
 import { PublicKey, Transaction } from '@solana/web3.js';
-import { PoolMetadata } from '../info';
 
 // Pool creation operations
 export interface CreatePairWithConfigParams {
@@ -34,23 +33,4 @@ export interface GetBinsReserveParams {
   position: PublicKey;
   pair: PublicKey;
   payer: PublicKey;
-}
-
-// Vault operations
-export interface GetUserVaultInfoParams {
-  tokenAddress: PublicKey;
-  payer: PublicKey;
-  transaction?: Transaction;
-}
-
-
-// Quote operations
-export interface QuoteParams {
-  amount: number;
-  metadata: PoolMetadata;
-  optional: {
-    isExactInput: boolean;
-    swapForY: boolean;
-    slippage: number;
-  };
 }
